requirements-local.txt
aiohttp>=3.8.0
pyautogui>=0.9.54
pillow>=10.0.0
pyperclip>=1.8.2
uiautomation>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
